Refactors useRegister into register.tsx for simplicity
