+++ lib/hooks/auth/index.ts
@@ -1,4 +1,3 @@
 export { useConfirm } from "./useConfirm";
 export { useLogin } from "./useLogin";
-export { useRegister } from "./useRegister";
 export { useResetPassword } from "./useResetPassword";
